requirements.txt
# 万相IP技能依赖包
# 版本要求基于Python 3.8+

# 核心依赖
requests>=2.28.0
Pillow>=9.0.0
python-dotenv>=0.19.0

# 可选依赖（用于高级功能）
# opencv-python>=4.5.0  # 图像处理
# numpy>=1.21.0  # 数值计算
# pandas>=1.3.0  # 数据处理

# 开发依赖
# pytest>=7.0.0  # 测试框架
# black>=22.0.0  # 代码格式化
# flake8>=4.0.0  # 代码检查